clothing_store/asgi.py
"""
ASGI config for clothing_store project.

It exposes the ASGI callable as a module-level variable named ``application``.

For more information on this file, see
https://docs.djangoproject.com/en/4.1/howto/deployment/asgi/
"""

import os
from decouple import config
from django.core.asgi import get_asgi_application

os.environ.setdefault(
    "DJANGO_SETTINGS_MODULE", f"clothing_store.settings.{config('SETTINGS')}"
)

application = get_asgi_application()
