requirements.txt
asgiref==3.7.2
black==23.11.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudinary==1.36.0
colorama==0.4.6
Django==4.2.6
django-autoslug==1.9.9
django-cloudinary-storage==0.3.0
django-debug-toolbar==4.2.0
django-jazzmin==2.6.0
idna==3.4
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==10.0.1
platformdirs==4.0.0
pluggy==1.3.0
psycopg==3.1.12
psycopg-binary==3.1.12
pytest==7.4.2
pytest-django==4.5.2
python-decouple==3.8
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
sweetify==2.3.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
whitenoise==6.6.0
